champion_detection/requirements.txt
# Core dependencies
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0

# CLIP model
clip-by-openai>=1.0
transformers>=4.30.0

# OpenAI API
openai>=1.0.0

# Image processing
scikit-image>=0.21.0
matplotlib>=3.7.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
python-dotenv>=1.0.0

# Data handling
pandas>=2.0.0
json5>=0.9.0

# Logging and monitoring
loguru>=0.7.0